fix: dokan vendor subscriptions inconsistencies (#825)
Issue 1: for Dokan Subscription product type, we are storing product_pack_enddate key data without checking _pack_validity (for non recurring) and _subscription_length (for recurring) subscriptions,

Issue 2: For all avaliable payment gateway in dokan pro, we are not checking if product type is product_pack (Dokan Subscription product) and overridding user_meta with any subscription type product. This was causing override of existing Dokan Subscription package user subscribed.

Issue 3: Issue 2 is causing problems with subscription product purchase, as we are storing user_meta for any subscription product type, subscriptions were getting canceled after initial payment.

*added migration process to update product_pack_enddate for existing active Dokan Subscription product type purchase.

#800

+++ modules/stripe/includes/WebhooksEvents/SubscriptionDeleted.php
@@ -43,6 +43,11 @@ class SubscriptionDeleted implements WebhookHandleable {
         $vendor_id     = Helper::get_vendor_id_by_subscription( $subscription->id );
         $order_id      = get_user_meta( $vendor_id, 'product_order_id', true );
         $has_recurring = get_user_meta( $vendor_id, '_customer_recurring_subscription', true );
+        $product_id    = get_user_meta( $vendor_id, 'product_package_id', true );
+
+        if ( ! SubscriptionHelper::is_subscription_product( $product_id ) ) {
+            return;
+        }
 
         if ( $has_recurring ) {
             SubscriptionHelper::delete_subscription_pack( $vendor_id, $order_id );
